add list animations
