+++ src/views/dashboard/components/episode-feed.vue
@@ -1,30 +1,10 @@
 <template>
   <div class="episode-feed">
-    <div class="episodes">
-      <router-link
-        v-for="schedule in airingSchedules"
-        :key="schedule.id"
-        :to="`/anime/${schedule.media.id}`"
-        class="episode"
-      >
-        <img
-          :src="schedule.media.coverImage.medium"
-          class="banner"
-          :style="{ backgroundColor: schedule.media.coverImage.color }"
-        />
-
-        <div class="info">
-          <div class="title">{{ schedule.media.title.userPreferred }}</div>
-
-          <next-episode-info
-            :nextAiringEpisode="{
-              episode: schedule.episode,
-              airingAt: schedule.airingAt,
-            }"
-            class="airing"
-          />
-        </div>
-      </router-link>
+    <div class="episodes-container">
+      <animated-list
+        v-if="listIds && queueIds"
+        :ids="mode === EpisodeFeedMode.LIST ? listIds : queueIds"
+      />
     </div>
 
     <div class="control-panel">
@@ -52,19 +32,14 @@
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator'
 import { oc } from 'ts-optchain'
-import { addDays as _addDays, startOfDay } from 'date-fns'
 import { mdiClipboardTextOutline, mdiPlaylistCheck } from '@mdi/js'
 
 import Icon from '@/common/components/icon.vue'
-import NextEpisodeInfo from '@/common/components/next-episode-info.vue'
-import EPISODE_FEED_QUERY from './episode-feed.graphql'
+import AnimatedList from './animated-list.vue'
 import LIST_IDS_QUERY from './episode-feed-list-ids.graphql'
 import {
-  EpisodeFeedAiringSchedules,
   EpisodeFeedListIdsQuery,
   EpisodeFeedListIdsVariables,
-  EpisodeFeedQuery,
-  EpisodeFeedVariables,
 } from '@/graphql/types'
 
 import { Query } from '@/decorators'
@@ -77,12 +52,7 @@ import { getAnilistUserId } from '@/state/auth'
 import { getQueue } from '@/state/user'
 import { prop } from '@/utils'
 
-// startOfDay is used to make sure we can cache the results properly
-// if we don't, every query will have different time variables
-const addDays = (days: number) =>
-  Math.round(startOfDay(_addDays(new Date(), days)).valueOf() / 1000)
-
-@Component({ components: { Icon, NextEpisodeInfo } })
+@Component({ components: { AnimatedList, Icon } })
 export default class EpisodeFeed extends Vue {
   @Query<EpisodeFeed, EpisodeFeedListIdsQuery, EpisodeFeedListIdsVariables>({
     query: LIST_IDS_QUERY,
@@ -100,27 +70,6 @@ export default class EpisodeFeed extends Vue {
     },
   })
   public listIds!: number[]
-
-  @Query<EpisodeFeed, EpisodeFeedQuery, EpisodeFeedVariables>({
-    fetchPolicy: 'cache-first',
-    query: EPISODE_FEED_QUERY,
-    variables() {
-      return {
-        page: this.page,
-        startDate: addDays(-2),
-        endDate: addDays(7),
-        ids: this.mode === EpisodeFeedMode.LIST ? this.listIds : this.queueIds,
-      }
-    },
-    update(data) {
-      this.hasNextPage = oc(data).Page.pageInfo.hasNextPage(false)
-
-      return oc(data).Page.airingSchedules(null)
-    },
-  })
-  public airingSchedules!: EpisodeFeedAiringSchedules
-  public hasNextPage = false
-  public page = 1
 
   public get userId() {
     return getAnilistUserId(this.$store)
@@ -155,8 +104,17 @@ export default class EpisodeFeed extends Vue {
   height: 100%;
   width: 100%;
 
+  & > .episodes-container {
+    position: relative;
+    width: 100%;
+    height: 100%;
+    overflow-y: scroll;
+    direction: rtl;
+  }
+
   & > .control-panel {
     flex-shrink: 0;
+    margin-top: auto;
     height: 45px;
     width: 100%;
     background: color($dark, 300);
@@ -180,65 +138,6 @@ export default class EpisodeFeed extends Vue {
       & > .icon {
         height: 30px;
         fill: $white;
-      }
-    }
-  }
-
-  & > .episodes {
-    height: 100%;
-    padding: 15px 0;
-
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: flex-start;
-    overflow: auto;
-    direction: rtl;
-
-    & .episode {
-      direction: ltr;
-      flex-shrink: 0;
-
-      width: 95%;
-      display: flex;
-      position: relative;
-      margin-bottom: 15px;
-      overflow: hidden;
-
-      background: $dark;
-      border-radius: 5px;
-      color: $white;
-      text-decoration: none;
-      box-shadow: $shadow;
-
-      &:last-child {
-        margin-bottom: 0;
-      }
-
-      & > .banner {
-        flex-shrink: 0;
-        height: 100%;
-        width: 60px;
-        object-fit: cover;
-      }
-
-      & > .info {
-        width: calc(100% - 60px);
-        position: relative;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        padding: 0 15px;
-        text-align: left;
-
-        & > .title {
-          font-family: 'Raleway', sans-serif;
-          font-weight: 500;
-          white-space: nowrap;
-          overflow: hidden;
-          text-overflow: ellipsis;
-          margin-bottom: 5px;
-        }
       }
     }
   }
